pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pydessem"
version = "0.1.0"
description = "Modelo didático tipo DESSEM em Pyomo (UC térmico, hidráulica PWL e reservas)."
readme = "README.md"
authors = [{name="Sr. (autor)", email="author@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
  "pyomo>=6.7",
  "pyyaml>=6.0.1",
  "numpy>=1.24"
]

[project.urls]
Homepage = "https://example.com/pydessem"

[project.scripts]
pydessem-solve = "pydessem.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
